requirements.txt
torch
torchvision
torchaudio
python-chess
tqdm
numpy
scikit-learn
pillow
matplotlib